Automatically backed up by Learn

+++ db/migrate/20181219150748_add_description_inventory_to_product.rb
@@ -1,6 +1,6 @@
 class AddDescriptionInventoryToProduct < ActiveRecord::Migration
   def change
     add_column :products, :description, :text
-    add_column :products, :inventory, :integer
+    add_column :products, :inventory, :decimal
   end
 end
